+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added support for `GlobalUniqueIndex` type in `table/options.IndexDescription`
+
 ## v3.125.3
 * Added optimized helper methods for efficient `[]byte` → `string` conversions in `JSON` and `JSONDocument` types.